Refactoring investor statistic card - moved to components/cards - moved InfoCard to common - clean code
