+++ src/common/CardInfo/styled.ts
