+++ src/common/index.ts
@@ -9,3 +9,4 @@ export { default as SelectableCard } from "./SelectableCard"
 export { default as StepsNavigation } from "./StepsNavigation"
 export { default as TokenChip } from "./TokenChip"
 export { default as DaoSettingsParameters } from "./DaoSettingsParameters"
+export { default as CardInfo } from "./CardInfo"
